Skip second factor if we cannot identify the user.
We configure the attribute that we're going to use to identify users in subsequent authentications. However, if that attribute is missing (for whatever reason), we shouldn't fail horribly with a blank page. Instead, we just skip the second factor (should we configure what behaviour we want here? Does it make sense to fail and abort authentication if we cannot perform 2-factor authentication?).